Api 474 fortify april 2021 (#725)
* Empty catch block

*Log error message instead of error

* Redundant null check in IncludesIcnMajig

* Header Manipulation

On IcnIncludesMajig:56, Fortify complains that the X_VA_INCLUDES_ICN header can be manipulated by an external source. To prevent this, a regex match was added to check the patientICN was valid.

* Log forging : sanitize the logs by replacing CRLF

* Obsolete

Fortify was complaining that the lombok generated methods were using a deprecated variable. Instead of revving data-query to the next major version, we double checked this var was not being used in any other applications and removed it.